fix: everything

+++ app/api/remove-bg/route.ts
@@ -1,10 +1,34 @@
-import { imagekit, replicate } from "@/lib/config";
+import { imagekit } from "@/lib/config";
 import { NextRequest, NextResponse } from "next/server";
+import { fal } from "@fal-ai/client";
+import prisma from "@/lib/prisma";
+import { auth } from "@clerk/nextjs/server";
 
+fal.config({
+    credentials: process.env.FAL_KEY!,
+});
 
-export async function POST(req: NextRequest) { 
+export async function POST(req: NextRequest) {
 
     try {
+        const { userId } = await auth();
+
+        if (!userId) {
+            return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+        }
+
+        const user = await prisma.user.findUnique({
+            where: { id: userId },
+            select: { credits: true },
+        });
+
+        if (!user || user.credits <= 0) {
+            return NextResponse.json(
+                { error: "Insufficient credits" },
+                { status: 403 }
+            );
+        }
+
         const formData = await req.formData();
         const file = formData.get("image") as File;
 
@@ -16,23 +40,35 @@ export async function POST(req: NextRequest) {
         const buffer = Buffer.from(bytes);
 
         const uploadResult = await imagekit.upload({
-            file: buffer.toString("base64"),
+            file: buffer,
             fileName: file.name,
             isPublished: true,
             useUniqueFileName: false,
         });
 
-        const input = {
-            image: uploadResult.url,
-        };
+        const result = await fal.subscribe("fal-ai/imageutils/rembg", {
+            input: {
+                image_url: uploadResult.url
+            },
+            logs: true,
+            onQueueUpdate: (update) => {
+                if (update.status === "IN_PROGRESS") {
+                    update.logs.map((log) => log.message).forEach(console.log);
+                }
+            },
+        });
 
-        const output = await replicate.run("lucataco/remove-bg:95fcc2a26d3899cd6c2691c900465aaeff466285a65c14638cc5f36f34befaf1", { input }) as unknown as string;
+        const imageUrl = result.data.image.url
+        if (!imageUrl) {
+            return NextResponse.json(
+                { error: "No image generated" },
+                { status: 500 }
+            )
+        }
 
-        console.log(output);
-        
         await imagekit.deleteFile(uploadResult.fileId);
 
-        const response = await fetch(output);
+        const response = await fetch(imageUrl);
         const arrayBuffer = await response.arrayBuffer();
         const base64 = Buffer.from(arrayBuffer).toString("base64");
 
@@ -42,7 +78,19 @@ export async function POST(req: NextRequest) {
             isPublished: true,
             useUniqueFileName: false,
         });
-        
+
+        await prisma.image.create({
+            data: {
+                url: upscaleUpload.url,
+                fileId: upscaleUpload.fileId,
+                userId: userId!
+            }
+        })
+
+        await prisma.user.update({
+            where: { id: userId },
+            data: { credits: { decrement: 1 } },
+        });
 
         return NextResponse.json({
             success: true,
@@ -50,7 +98,7 @@ export async function POST(req: NextRequest) {
             msg: "Image upscaled and stored successfully",
         });
     } catch (error: any) {
-        console.log(error.message,"Message");
+        console.log(error.message, "Message");
         return NextResponse.json(
             { error: error.message || "Failed to upscale image" },
             { status: 500 }
